requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
boto3==1.34.121
botocore==1.34.121
cachecontrol==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.8
multidict==6.0.5
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.3
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twilio==9.0.2
twisted==24.3.0
#twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
yarl==1.9.4
zope.interface==6.2
